website: update anchor tags for external pages to open in a new window
When following the steps in the contributor guide, it is easy to lose your
place in the list when every link opens in the same tab. By having external pages
open in new tabs, it will be easier to close those tabs as steps are completed and
go back to the original list in the original tab.